Regional form tests

+++ include/pokemon.h
@@ -911,7 +911,7 @@ const u8 *GetMoveAnimationScript(u16 moveId);
 void UpdateDaysPassedSinceFormChange(u16 days);
 void TrySetDayLimitToFormChange(struct Pokemon *mon);
 u32 CheckDynamicMoveType(struct Pokemon *mon, u32 move, u32 battler);
-u16 GetRegionalForm(u32 species, u32 region);
+u32 GetRegionalForm(u32 species, u32 region);
 bool32 IsSpeciesForeignRegionalForm(u32 species, u32 currentRegion);
 
 #endif // GUARD_POKEMON_H
